Breakup Det into many LL modules
